more code
Now we have everything for running some basic packet forwarding.

+++ miki/packet.h
@@ -15,19 +15,21 @@
 
 struct Interface;
 
+//TODO how can we mark something as payload?
+//  use type=0? no, we need typed payload
 struct PacketHeader {
     int type;
-    off_t start; // from beginning of buf
-    size_t len;
+    unsigned start; // from beginning of buf
+    unsigned len;
 };
 
 struct Packet {
     unsigned char *buf;
-    off_t start; // from beginning of buf
-    size_t len;
+    unsigned start; // from beginning of buf
+    unsigned len;
 
     // scratch space starts at offset 0
-    size_t scratch_len;
+    unsigned scratch_len;
 
     struct PacketHeader headers[PACKET_MAX_HEADER_NUM];
     unsigned header_count;
@@ -54,14 +56,14 @@ bool packet_dummy(const struct Packet *p);
 // identify the header at the given position in the packet as @type
 // this will create a new entry in p->headers
 // @offset is counted from p->start
-// the entries of p->headers will be ordered by their offset
+// the headers MUST be identified by increasing offset
 // TODO do we prevent the headers from overlapping? do we have to?
-void packet_identify_header(struct Packet *p, int type, off_t offset, size_t len);
+void packet_identify_header(struct Packet *p, int type, unsigned offset, unsigned len);
 
 // adds a new header on the scratch space
 // the position in the header list is @idx
 // all the existing headers after @idx will be shifted in the array
-void packet_add_header(struct Packet *p, unsigned idx, int type, size_t len);
+void packet_add_header(struct Packet *p, unsigned idx, int type, unsigned len);
 
 // removes a header and forgets it
 // all the headers after @idx will be shifted in the array
